requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pypdf
nltk
sentence-transformers
faiss-cpu